Update ns-3 to ns-3-dev

+++ src/applications/model/udp-client.h
@@ -94,7 +94,7 @@ private:
   Time m_interval; //!< Packet inter-send time
   uint32_t m_size; //!< Size of the sent packet (including the SeqTsHeader)
 
-  uint32_t m_sent; //!< Counter for sent packets
+  uint64_t m_sent; //!< Counter for sent packets
   Ptr<Socket> m_socket; //!< Socket
   Address m_peerAddress; //!< Remote peer address
   uint16_t m_peerPort; //!< Remote peer port
