+++ src/applications/model/udp-server.h
@@ -105,7 +105,7 @@ private:
   uint16_t m_port; //!< Port on which we listen for incoming packets.
   Ptr<Socket> m_socket; //!< IPv4 Socket
   Ptr<Socket> m_socket6; //!< IPv6 Socket
-  uint32_t m_received; //!< Number of received packets
+  uint64_t m_received; //!< Number of received packets
   PacketLossCounter m_lossCounter; //!< Lost packet counter
 
   std::string m_receivedFilename;
